+++ src/lib/storage/factory.cpp
@@ -1,6 +1,6 @@
 /*
  -------------------------------------------------------------------------------
-    This file is part of the thermos.
+    This file is part of thermos.
     Copyright (C) 2022  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
@@ -18,17 +18,25 @@
  -------------------------------------------------------------------------------
 */
 
-#ifndef THERMOS_VERSION_HPP
-#define THERMOS_VERSION_HPP
+#include "factory.hpp"
+#include "csv.hpp"
+#include "db.hpp"
 
-#include <string>
-
-namespace thermos
+namespace thermos::storage
 {
 
-/** \brief version information */
-const std::string version = "version 0.3.0-pre, 2022-04-29";
+std::unique_ptr<store> factory::create(const type t)
+{
+  switch (t)
+  {
+    case type::db:
+         return std::make_unique<db>();
+    case type::csv:
+         return std::make_unique<csv>();
+    default:
+         // Any future unsupported type returns a null pointer.
+         return nullptr;
+  }
+}
 
 } // namespace
-
-#endif // THERMOS_VERSION_HPP