+++ changelog.md
@@ -1,5 +1,14 @@
 # Version history of thermos
 
+## Version 0.3.0-pre (2022-04-29)
+
+`thermo-logger` can now alternatively write the logged data into a SQLite 3
+database. The file format can be set by the new `--type` parameter. While
+`--type csv` still writes data to a CSV file, specifying '--type db` will cause
+the program to write the data as SQLite 3 database instead.
+
+If no type is given, then the file is assumed to be a SQLite 3 database.
+
 ## Version 0.2.0 (2022-04-24)
 
 `thermo-logger`, a command line application that continuously logs thermal